teste de integracao finalizado e passando

+++ core/src/test/java/br/com/gubee/interview/core/features/hero/HeroControllerTestEndpointIT.java
@@ -1,15 +1,14 @@
 package br.com.gubee.interview.core.features.hero;
 
+import br.com.gubee.interview.core.exception.NotFoundException;
 import br.com.gubee.interview.core.features.powerstats.PowerStatsRepository;
 import br.com.gubee.interview.core.features.util.TestFactory;
-import br.com.gubee.interview.model.enums.Race;
 import br.com.gubee.interview.model.hero.Hero;
 import br.com.gubee.interview.model.hero.dto.HeroRequest;
 import br.com.gubee.interview.model.hero.dto.HeroResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -19,14 +18,15 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import java.util.List;
 import java.util.UUID;
 import java.util.function.Supplier;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -37,7 +37,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 class HeroControllerTestEndpointIT {
 
     @Container
-    static final PostgreSQLContainer postgresContainer = new PostgreSQLContainer("postgres:12");
+    static final PostgreSQLContainer<?> postgresContainer = new PostgreSQLContainer<>("postgres:12")
+            .withReuse(false);
+
     @Autowired
     private MockMvc mockMvc;
     @Autowired
@@ -46,14 +48,16 @@ class HeroControllerTestEndpointIT {
     private PowerStatsRepository powerStatsRepository;
     @Autowired
     private HeroRepository heroRepository;
-    private  final String BASE_PATH = "/api/v1/heroes";
+    private final String BASE_PATH = "/api/v1/heroes";
     UUID powerStatsId;
     UUID heroId;
 
     @AfterEach
     void afterEach() {
         if (powerStatsId != null) {
-            heroRepository.deleteById(heroId);
+            if (heroId != null) {
+                heroRepository.deleteById(heroId);
+            }
             powerStatsRepository.deleteById(powerStatsId);
         }
     }
@@ -62,9 +66,9 @@ class HeroControllerTestEndpointIT {
     void create() throws Exception {
         powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
         String body = objectMapper.writeValueAsString(TestFactory.createHeroRequest(powerStatsId));
-        ResultActions actions = perform(body, () -> MockMvcRequestBuilders.post(BASE_PATH));
+        ResultActions actions = perform(body, () -> post(BASE_PATH));
         actions.andExpect(status().isCreated()).andExpect(header().exists("Location"));
-        heroId  = toUUID(actions);
+        heroId = toUUID(actions);
         Hero hero = heroRepository.findById(heroId);
         Assertions.assertEquals(heroId, hero.getId());
     }
@@ -72,7 +76,7 @@ class HeroControllerTestEndpointIT {
     @Test
     void createNotFoundPowerStatsId() throws Exception {
         String body = objectMapper.writeValueAsString(TestFactory.createHeroRequest());
-        ResultActions actions = perform(body, () -> MockMvcRequestBuilders.post(BASE_PATH));
+        ResultActions actions = perform(body, () -> post(BASE_PATH));
         actions.andExpect(status().isNotFound());
     }
 
@@ -80,52 +84,89 @@ class HeroControllerTestEndpointIT {
     void update() throws Exception {
         //  given
         powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
-        Hero flash = TestFactory.createHero(powerStatsId);
-        heroId = heroRepository.create(TestFactory.createHero(powerStatsId));
+        HeroRequest flash = TestFactory.createHeroRequest(powerStatsId);
+        heroId = heroRepository.create(HeroMapper.toHero(flash));
         //  when
-        flash.setRace(Race.DIVINE);
-        ResultActions actions = perform(objectMapper.writeValueAsString(flash), () -> put(BASE_PATH+"/{id}", flash.getId()));
+        ResultActions actions = perform(
+                objectMapper.writeValueAsString(TestFactory.createHeroRequest("Superman", powerStatsId))
+                , () -> put(BASE_PATH + "/{id}", heroId)
+        );
         //  then
+        Hero hero = heroRepository.findById(heroId);
         actions.andExpect(status().isOk());
-        Assertions.assertEquals(Race.DIVINE, heroRepository.findById(heroId).getRace());
+        Assertions.assertEquals("Superman", hero.getName());
     }
 
 
     @Test
     void updateHeroIdNotFound() throws Exception {
-        //  when
-        ResultActions actions = perform(objectMapper.writeValueAsString(TestFactory.createHeroRequest()), () -> put(BASE_PATH+ "/{id}", UUID.randomUUID()));
-        //  then
+        ResultActions actions = perform(objectMapper.writeValueAsString(TestFactory.createHeroRequest()), () -> put(BASE_PATH + "/{id}", UUID.randomUUID()));
         actions.andExpect(status().isNotFound());
     }
 
     @Test
     void updatePowerStatsIdNotFound() throws Exception {
-        Hero hero = heroRepository.findAll().get(0);
-        HeroRequest request = TestFactory.createHeroRequest();
+        //  given
+        powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
+        var request = TestFactory.createHeroRequest(powerStatsId);
+        heroId = heroRepository.create(HeroMapper.toHero(request));
+        //  when
         request.setStats(UUID.randomUUID());
-        request.setRace(Race.CYBORG);
-        ResultActions actions = perform(objectMapper.writeValueAsString(request), () -> put(BASE_PATH+"/{id}", hero.getId()));
+        //  then
+        ResultActions actions = perform(objectMapper.writeValueAsString(request), () -> put(BASE_PATH + "/{id}", heroId));
         actions.andExpect(status().isNotFound());
     }
 
     @Test
     void getHeroByID() throws Exception {
-        Hero hero = heroRepository.findAll().get(0);
-        ResultActions actions = perform(objectMapper.writeValueAsString(hero), () -> get(BASE_PATH+"/{id}", hero.getId()));
+        powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
+        var request = TestFactory.createHeroRequest(powerStatsId);
+        heroId = heroRepository.create(HeroMapper.toHero(request));
+        ResultActions actions = perform(() -> get(BASE_PATH + "/{id}", heroId));
         String content = actions.andExpect(status().isOk()).andReturn().getResponse().getContentAsString();
-        HeroResponse response  = objectMapper.readValue(content, HeroResponse.class);
-        Assertions.assertEquals(hero.getName(), response.getName());
-        Assertions.assertEquals(hero.getRace(), response.getRace());
-        Assertions.assertEquals(hero.getPowerStatsId(), response.getPowerStatsId());
+        HeroResponse response = objectMapper.readValue(content, HeroResponse.class);
+        Assertions.assertEquals(request.getName(), response.getName());
+        Assertions.assertEquals(request.getRace(), response.getRace());
+        Assertions.assertEquals(request.getStats(), response.getPowerStatsId());
     }
 
     @Test
-    void deleteById() {
+    void deleteById() throws Exception {
+        powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
+        var request = TestFactory.createHeroRequest(powerStatsId);
+        heroId = heroRepository.create(HeroMapper.toHero(request));
+        ResultActions actions = perform(() -> delete(BASE_PATH + "/{id}", heroId));
+        actions.andExpect(status().isOk());
+        Assertions.assertThrows(NotFoundException.class, () -> {
+           heroRepository.findById(heroId);
+        });
+        heroId = null;
     }
 
     @Test
-    void getHeroByNameLike() {
+    void deleteHeroIdNotFound() throws Exception {
+        ResultActions actions = perform(() -> delete(BASE_PATH + "/{id}", UUID.randomUUID()));
+        actions.andExpect(status().isNotFound());
+    }
+
+    @Test
+    void getHeroByNameLike() throws Exception {
+        //  given
+        powerStatsId = powerStatsRepository.create(TestFactory.createPowerStats());
+        List<Hero> heroes = Stream.of("vanderson", "evandro", "ivan")
+                .map(name -> TestFactory.createHeroRequest(name, powerStatsId))
+                .map(HeroMapper::toHero)
+                .collect(Collectors.toList());
+        List<UUID> ids = heroes.stream().map(hero -> heroRepository.create(hero)).collect(Collectors.toList());
+        //  when
+        ResultActions actions = perform(() -> get(BASE_PATH+"/search/van"));
+        //  then
+        actions.andExpect(status().isOk());
+        String contentAsString = actions.andReturn().getResponse().getContentAsString();
+        HeroResponse[] heroResponses = objectMapper.readValue(contentAsString, HeroResponse[].class);
+        Assertions.assertEquals(3, heroResponses.length);
+        //  Com assertJ -> assertThat(heroResponses).hasSize(3);
+        ids.forEach(id -> heroRepository.deleteById(id));
     }
 
     ResultActions perform(String body, Supplier<MockHttpServletRequestBuilder> method)
@@ -144,7 +185,7 @@ class HeroControllerTestEndpointIT {
     UUID toUUID(ResultActions actions) {
         var uri = actions.andReturn().getResponse().getHeader("Location");
         if (uri == null) throw new RuntimeException();
-        return UUID.fromString( uri.substring(uri.lastIndexOf("/") + 1));
+        return UUID.fromString(uri.substring(uri.lastIndexOf("/") + 1));
     }
